requirements.txt
aiogram>=3.13,<4
telethon>=1.36,<2
python-dotenv>=1.0,<2
aiofiles>=24.1,<25
